add levels to the json file
Wanted a way to pass in the level labels without having to change
the source code.  Creating a separate object in the json file and
passing the values from that.  If the levels are not specified in
the json file, then use the default values that were in the code.